Update seadatacloud-nvs2.sh
Update to match the changes of the script. 
NVS new launch

We are very pleased to announce that we have launched NVS with new functionalities and "look and feel".
So please have a look at http://vocab.nerc.ac.uk

For anybody that was accessing the RDF/XML pages, you will get them back if you change your Accept header to application/rdf+xml or by clicking on the Alternate Profiles--> RDF/XML
NOTE: Currently the html interface does not show all the mappings although they exist, so we are very sorry about that. 
This will be corrected on Monday. In the mean time, to see your mappings please click on the Alternate Profiles--> RDF/XML or any other serialization you want. 
We would love to receive your feedback